Deletes profile sub folder upon deletion of last file.

+++ src/database/subClasses/FileMethods.java
@@ -23,7 +23,6 @@ public class FileMethods {
     private FilePathGenerator fpg;
     private ExperimentMethods expMethods;
 
-
     /**
      * Constructor for the fileMethod object.
      *
@@ -36,7 +35,6 @@ public class FileMethods {
         fpg = filePG;
         this.expMethods = expMethods;
     }
-
 
     /**
      * @param expID
@@ -109,18 +107,18 @@ public class FileMethods {
         stmt.setString(1, path);
 
         switch (fileType) {
-            case FileTuple.RAW:
-                stmt.setString(2, "Raw");
-                break;
-            case FileTuple.PROFILE:
-                stmt.setString(2, "Profile");
-                break;
-            case FileTuple.REGION:
-                stmt.setString(2, "Region");
-                break;
-            default:
-                stmt.setString(2, "Other");
-                break;
+        case FileTuple.RAW:
+            stmt.setString(2, "Raw");
+            break;
+        case FileTuple.PROFILE:
+            stmt.setString(2, "Profile");
+            break;
+        case FileTuple.REGION:
+            stmt.setString(2, "Region");
+            break;
+        default:
+            stmt.setString(2, "Other");
+            break;
         }
 
         stmt.setString(3, fileName);
@@ -138,7 +136,6 @@ public class FileMethods {
         return getFileTuple(path);
     }
 
-
     private FileTuple getProfile(Experiment e, String metaData) {
         for (FileTuple ft : e.getFiles()) {
             if (ft.type.equalsIgnoreCase("profile")
@@ -149,7 +146,6 @@ public class FileMethods {
         return null;
     }
 
-
     private String getParentFolder(String filePath) {
         int filenameIndex = filePath.lastIndexOf(File.separator);
         return filePath.substring(0, filenameIndex + 1);
@@ -159,7 +155,6 @@ public class FileMethods {
         int filenameIndex = filePath.lastIndexOf(File.separator);
         return filePath.substring(filenameIndex + 1);
     }
-
 
     /**
      * Returns the FileTuple object associated with the given filePath.
@@ -186,7 +181,6 @@ public class FileMethods {
         return null;
     }
 
-
     /**
      * Returns the FileTuple object associated with the given fileID.
      *
@@ -212,7 +206,6 @@ public class FileMethods {
         return null;
     }
 
-
     /**
      * Deletes a file from the database and the disk. Should throw an
      * IOException if the method failed to delete the file from disk.
@@ -222,11 +215,24 @@ public class FileMethods {
      * @return int - the number of deleted tuples in the database.
      * @throws SQLException
      *             if the query does not succeed
+     * @throws IOException
      */
-    public int deleteFile(String path) throws SQLException {
+    public int deleteFile(String path) throws SQLException, IOException {
+
         File fileToDelete = new File(path);
         if (fileToDelete.exists()) {
             fileToDelete.delete();
+        }
+
+        FileTuple ft = getFileTuple(path);
+
+        if (ft == null) {
+            throw new IOException("Could not find file at path " + path);
+        }
+
+        File parentFolder = new File(ft.getParentFolder());
+        if (ft.type.equalsIgnoreCase("profile") && isEmptyFolder(parentFolder)) {
+            parentFolder.delete();
         }
 
         String statementStr = "DELETE FROM File " + "WHERE (Path ~~* ?)";
@@ -238,38 +244,66 @@ public class FileMethods {
         return resCount;
     }
 
-
     /**
-     * Deletes a file from the database and the disk using the fileID. Should
-     * throw an IOException if the method failed to delete the file from disk.
+     * Deletes a file from the database and the disk using the fileID.
      *
      * @param fileID
      *            int - the fileID of the file to be deleted.
      * @return 1 if deletion was successful, else 0.
-     * @throws SQLException
+     * @throws SQLException If the database could not be contacted
+     * @throws IOException If the FileID does not exist in the database
      */
-    public int deleteFile(int fileID) throws SQLException {
+    public int deleteFile(int fileID) throws SQLException, IOException {
 
-        String query1 = "SELECT Path FROM File " + "WHERE FileID = ?";
-        String query2 = "DELETE FROM File " + "WHERE FileID = ?";
-        int res = 0;
+        FileTuple ft = getFileTuple(fileID);
 
-        PreparedStatement stmt = conn.prepareStatement(query1);
-        stmt.setInt(1, fileID);
-        ResultSet rs = stmt.executeQuery();
-        if (rs.next()) {
-            File fileToDelete = new File(rs.getString("Path"));
-            if (fileToDelete.exists()) {
-                fileToDelete.delete();
-            }
-            stmt = conn.prepareStatement(query2);
-            stmt.setInt(1, fileID);
-            res = stmt.executeUpdate();
-            stmt.close();
+        if (ft == null) {
+            throw new IOException("Could not find file with ID " + fileID);
         }
+
+        File fileToDelete = new File(ft.path);
+
+        if (fileToDelete.exists()) {
+
+            fileToDelete.delete();
+
+            File parentFolder = new File(ft.getParentFolder());
+            if (ft.type.equalsIgnoreCase("profile") && isEmptyFolder(parentFolder)) {
+                parentFolder.delete();
+            }
+        }
+
+        String query = "DELETE FROM File " + "WHERE FileID = ?";
+        PreparedStatement stmt = conn.prepareStatement(query);
+        stmt.setInt(1, fileID);
+        int res = stmt.executeUpdate();
+        stmt.close();
         return res;
     }
 
+    private boolean isEmptyFolder(File f) {
+
+        if (f.exists() && f.listFiles() != null && f.listFiles().length == 0) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * Recursively deletes a folder with all it's subfolders and files.
+     * @param folder the folder to delete.
+     */
+    public void recursiveDelete(File folder) {
+        File[] contents = folder.listFiles();
+        if (contents == null || contents.length == 0) {
+            folder.delete();
+        } else {
+            for (File f : contents) {
+                recursiveDelete(f);
+            }
+        }
+        folder.delete();
+    }
 
     /**
      * Checks if the file with the specified fileID exists in the database.
@@ -297,7 +331,6 @@ public class FileMethods {
         return hasResult;
     }
 
-
     public int fileReadyForDownload(int fileID) throws SQLException {
 
         String statusUpdateString = "UPDATE File SET Status = 'Done' "
@@ -312,7 +345,6 @@ public class FileMethods {
 
         return resCount;
     }
-
 
     /**
      * Changes the Filename for a specific file with given fileID. This method
@@ -378,10 +410,10 @@ public class FileMethods {
         return resCount;
     }
 
-
     public FileTuple addGeneratedFile(String expId, int fileType,
             String filePath, String inputFileName, String metaData,
-            String uploader, boolean isPrivate, String grVersion) throws SQLException, IOException {
+            String uploader, boolean isPrivate, String grVersion)
+            throws SQLException, IOException {
 
         Experiment e = expMethods.getExperiment(expId);
         if (e == null) {
@@ -403,18 +435,18 @@ public class FileMethods {
         stmt.setString(1, filePath);
 
         switch (fileType) {
-            case FileTuple.RAW:
-                stmt.setString(2, "Raw");
-                break;
-            case FileTuple.PROFILE:
-                stmt.setString(2, "Profile");
-                break;
-            case FileTuple.REGION:
-                stmt.setString(2, "Region");
-                break;
-            default:
-                stmt.setString(2, "Other");
-                break;
+        case FileTuple.RAW:
+            stmt.setString(2, "Raw");
+            break;
+        case FileTuple.PROFILE:
+            stmt.setString(2, "Profile");
+            break;
+        case FileTuple.REGION:
+            stmt.setString(2, "Region");
+            break;
+        default:
+            stmt.setString(2, "Other");
+            break;
         }
 
         stmt.setString(3, getFileName(filePath));
